Move getBestByString to Utils Get Context from ViewHolder instead of using the fragment
