+++ app/src/main/java/app/wefridge/wefridge/ItemRecyclerViewAdapter.kt
@@ -1,22 +1,16 @@
 package app.wefridge.wefridge
 
-import android.content.Context
 import android.os.Bundle
-import android.text.format.DateFormat
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.annotation.IdRes
-import androidx.fragment.app.Fragment
 import androidx.navigation.findNavController
 import androidx.recyclerview.widget.RecyclerView
 import app.wefridge.wefridge.databinding.FragmentPantryBinding
 import app.wefridge.wefridge.model.Item
-import java.util.*
-import java.util.concurrent.TimeUnit
-import kotlin.math.abs
 
 const val ARG_MODEL = "model"
 /**
@@ -25,8 +19,7 @@ const val ARG_MODEL = "model"
  */
 class ItemRecyclerViewAdapter(
     private val values: List<Item>,
-    @IdRes private val clickAction: Int? = null,
-    val fragment: Fragment
+    @IdRes private val clickAction: Int? = null
 ) : RecyclerView.Adapter<ItemRecyclerViewAdapter.ViewHolder>() {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
@@ -55,29 +48,8 @@ class ItemRecyclerViewAdapter(
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
         val item = values[position]
         holder.contentView.text = item.name
-        holder.bestByView.text = getBestByString(item.bestByDate)//"Best by in ${item.bestByDate} days"
+        holder.bestByView.text = getBestByString(item.bestByDate, holder.itemView.context)
         holder.sharedIcon.visibility = if (item.isShared) View.VISIBLE else View.INVISIBLE
-    }
-
-    private fun getBestByString(best_by: Date?): String {
-        val ctx = fragment.activity?.applicationContext!!
-        val today = Date()
-        if (best_by == null) {
-            return ""
-        }
-        if (today > best_by) {
-            val dateFormat = DateFormat.getDateFormat(ctx)
-            return ctx.getString(R.string.best_by_overdue, dateFormat.format(best_by))
-        }
-        val differenceInDays = TimeUnit.DAYS.convert(abs(best_by.time - today.time), TimeUnit.MILLISECONDS)
-        if (differenceInDays == 0L) {
-            return ctx.getString(R.string.best_by_today)
-        }
-        return if (differenceInDays > 1) {
-            ctx.getString(R.string.best_by_plural, differenceInDays)
-        } else {
-            ctx.getString(R.string.best_by_singular)
-        }
     }
 
     override fun getItemCount(): Int = values.size
